+++ order/src/main/java/store/_0982/order/application/order/dto/OrderRegisterCommand.java
@@ -1,4 +1,4 @@
-package store._0982.order.application.dto;
+package store._0982.order.application.order.dto;
 
 import java.util.UUID;
